+++ packages/app/services/pod-handling/pod-proxy.js
@@ -1,15 +1,7 @@
 const { Readable } = require('stream');
 const { FormData } = require('formdata-node');
 const { FormDataEncoder } = require('form-data-encoder');
-
-function stream2buffer(stream) {
-  return new Promise((resolve, reject) => {
-    const _buf = [];
-    stream.on('data', chunk => _buf.push(chunk));
-    stream.on('end', () => resolve(Buffer.concat(_buf)));
-    stream.on('error', err => reject(err));
-  });
-}
+const { stream2buffer } = require('../../utils');
 
 module.exports = {
   name: 'pod-proxy',
